Fix Compile

+++ src/component/rank/rank_board_svr/logic/rank_manager.h
@@ -122,7 +122,7 @@ class rank_manager : public atfw::util::design_pattern::singleton<rank_manager> 
                                                uint64_t main_node);
 
   ATFW_EXPLICIT_NODISCARD_ATTR rpc::result_code_type upgrade_rank_to_main(
-      rpc::context& ctx, const PROJECT_NAMESPACE_ID::DRankKey& rank_key, int32_t db_router_version);
+      rpc::context& ctx, const PROJECT_NAMESPACE_ID::DRankKey& rank_key, uint64_t db_router_version);
   ATFW_EXPLICIT_NODISCARD_ATTR static rpc::result_code_type check_slave_and_highest_data_version_slave(
       rpc::context& ctx, const PROJECT_NAMESPACE_ID::DRankKey& rank_key,
       const PROJECT_NAMESPACE_ID::table_rank_router& new_db_router, std::pair<uint64_t, int64_t>& highest_slave_node);
